Change strategy for storing local variables.
Now all variables that can refer to heap objects
are stored on the root stack.
All arguments are also placed to the root stack.

Do not add Label() as the first instruction of function.
Label will be added only after converting to string
form.
This is done to simplify processing of instructions.
Otherwise we always need to consider a special case
for the first instruction.